--- components/ShoppingMallReview/ShoppingMallReview.stories.tsx
@@ -1,8 +1,0 @@
-import ShoppingMallReview from './ShoppingMallReview';
-
-export default {
-  title: 'Components/ShoppingMallReview',
-  component: ShoppingMallReview,
-};
-
-export const Basic = () => <ShoppingMallReview />;
